% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_reg}
\alias{create_reg}
\title{From a registration file, get registrant totals for a given date}
\usage{
create_reg(regdate, regfile)
}
\value{
A dataframe of registrants per year and county, 2010-2016 for the MAY DATA
}
\description{
From a registration file, get registrant totals for a given date
}
\examples{
yearXcounty_reg_SoS()

}
