% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{turnout_calc}
\alias{turnout_calc}
\title{Get turnout, registered voters, and total votes per vounty}
\usage{
turnout_calc(regfile, histfile, el_date, county)
}
\arguments{
\item{regfile}{total registration file, with extra variable for year collected named FILE_YEAR}

\item{histfile}{complete history file across all reg file years}

\item{county}{a vector of all counties in the state}

\item{el_dat}{the election date, in the Date format}
}
\value{
A dataframe with turnout, registered voters, and turnout per county
}
\description{
Get turnout, registered voters, and total votes per vounty
}
