% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{turnouts_county_graph}
\alias{turnouts_county_graph}
\title{From the list output of multiple turnout_calc()s, get county level data for graphind}
\usage{
turnouts_county_graph(turnout_list)
}
\arguments{
\item{regfile}{total registration file, with extra variable for year collected named FILE_YEAR}

\item{turnouts_list}{a list created using the turnout_calc() function described previously}
}
\description{
From the list output of multiple turnout_calc()s, get county level data for graphind
}
\examples{
turnouts_county_graph(list_of_Maine_turnouts)

}
