% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_table}
\alias{var_table}
\title{Get turnout, registered voters, and total votes per vounty}
\usage{
var_table(regfile, var_location, var_name)
}
\arguments{
\item{regfile}{total registration file, with extra variable for year collected named FILE_YEAR}

\item{var_location}{Numerical location of the variable of interest}

\item{var_name}{The name that is desired of the final variable}
}
\value{
A dataframe with the variable of interest for each county, each year
}
\description{
Get turnout, registered voters, and total votes per vounty
}
\examples{
var_table(MAINE_12_to_17_REG_FILE, 3, COUNTY)

}
