% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yearXcounty_reg}
\alias{yearXcounty_reg}
\title{Get total registrants per year per county}
\usage{
yearXcounty_reg(filesource)
}
\value{
A dataframe of registrants per year and county, 2012-2016
}
\description{
Get total registrants per year per county
}
\examples{
yearXcounty_reg()
}
