% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yearXcounty_reg_2017}
\alias{yearXcounty_reg_2017}
\title{Get total registrants per year per county MAY DATA}
\usage{
yearXcounty_reg_2017()
}
\value{
A dataframe of registrants per year and county, 2010-2016 for the MAY DATA
}
\description{
Get total registrants per year per county MAY DATA
}
\examples{
yearXcounty_reg_SoS()
}
