% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{counties}
\alias{counties}
\title{Get Counties for a State out of Clean Reg Files}
\usage{
counties(filesource)
}
\arguments{
\item{filesource}{Registration File location}
}
\value{
A character vector of counties
}
\description{
Get Counties for a State out of Clean Reg Files
}
\examples{
counties(reg_colorado_16)
}
